pyproject.toml
[build-system]
requires = ["setuptools>=65.5.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sustainable-gh-workflow-linter"
version = "0.0.0"
description = "A linter that makes GitHub workflows more sustainable."
authors = [{name = "Kasper van Maasdam"}, {name = "Rebecca Andrei"}, {name = "Serban Mihai-Radu"}]
readme = "README.md"
requires-python = ">=3.11"
license = "GPL-3.0-or-later"
license-files = ["LICENSE"]
dependencies = [
    "ruamel.yaml>=0.18.14,<1.0.0"
]

[project.optional-dependencies]
dev = [
    "mypy>=1.19.0,<2.0.0",
    "pytest>=8.4.0,<9.0.0"
]

[project.scripts]
suslint = "suslint.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["suslint*"]


[tool.mypy]
ignore_missing_imports = true
warn_unused_configs = true
strict = true
check_untyped_defs = true

disallow_untyped_calls = false
disallow_untyped_defs = false
warn_unused_ignores = true
warn_redundant_casts = true
